requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
streamlit
pandas
plotly
python-dateutil
watchdog
extra-streamlit-components
